feat(react-middleware): add react-middleware package
