+++ .catrc.js
@@ -80,8 +80,11 @@ const babel = config => {
       plugins: [['@babel/proposal-pipeline-operator', { proposal: 'minimal' }]],
     },
     {
-      test: './server/server',
+      test: './server/react-middleware',
       presets: ['@babel/preset-react'],
+    },
+    {
+      test: './server/server',
       plugins: [['@babel/proposal-pipeline-operator', { proposal: 'minimal' }]],
     },
   );
