+++ server/react-middleware/.npmignore
@@ -1,0 +1,2 @@
+# babel
+src
