+++ server/react-middleware/README.md
@@ -1,0 +1,17 @@
+# [@cat-org/react-middleware][website] · <!-- badges.start -->[![npm][npm-image]][npm-link] ![npm-size][npm-size-image]
+
+[npm-image]: https://img.shields.io/npm/v/@cat-org/react-middleware.svg
+[npm-link]: https://www.npmjs.com/package/@cat-org/react-middleware
+[npm-size-image]: https://img.shields.io/bundlephobia/minzip/@cat-org/react-middleware.svg
+
+<!-- badges.end -->
+
+[website]: https://cat-org.github.io/core/react-middleware
+
+Cat react-middleware.
+
+## Install
+
+```sh
+yarn add @cat-org/react-middleware
+```
